outlook_message_rule: handle sequence
Sequence is ignored but required during creation. During creation,
service only honor the creation order and assign the actual sequence
based on that. In order to make the specified sequence match the final
one, one option is to addtionally specify `depends_on`, as is done in
acctest.

Besides, the API will return 500 if message rule is absent (rather than
404).